ksh-aliases: complain_missing() renamed to complain_noexec().
I like the structure of letting 'if' and 'elif' represent reasons to
bail the procedure of assigning a particular set of aliases and leaving
'else' as the successful path. Since there are often many aliases, it
makes more sense to have them at the bottom and the error conditions
first so the reader doesn't have to scroll past them to see the overall
logic.