remove filter_type from set options (explanation below).
libswresample appears to use kaiser as the default filter_type. I
produced spectrograms with sox(1) for a sample rate conversion from
44.1kHz to 48kHz with ffmpeg(1) using different filter_type arguments.
the spectrograms of blackman_nuttall and kaiser files
differed--cooler colors at the highest frequencies on kaiser's
side--but kaiser and default were the same (and diff(1) confirmed
this on both the FLAC and PNG files). the kaiser window approximates
the DPSS window, a window which maximizes energy concentration in the
main lobe, but is difficult to compute. kaiser is used in AAC.