rffd: change design for maintainability reasons
Have gone through and refactored most of the code so that it's more
general.

Rather than having a separate subroutine for each file category's
File::Find callback, all of them can be handled in a simple if-else
due to this refactor.

It did increase the complexity of the program's source code a little bit
(it now uses separate subroutines for many tasks, such as generating
file extensions, patterns to match extensions and mimetypes, etc). In
return, though, everything is abstracted and much more easily revised.

It also appears to be faster to keep the compiled regular expressions in
a hash.