sysinfo: fix a bug with battery charge display
100% charge was never displayed because the battery_hours_and_minutes
subroutine had called check_num_return_int beforehand, which warns and
returns if called against something that isn't a number. `apm -m`
displays unknown if the battery is fully charged, so the code should
handle that case.

While here, remove the early check on the capturex in
battery_hours_and_minutes. This gets checked later anyway and raises
an unnecessary warning in the case of 100% battery charge.