dmenuunicode: do a better job validating input
I did some testing with zzuf and it makes sense from a security
perspective to validate everything _before_ sending it to dmenu. If an
attacker managed to evade the basic checking from before, they could in
theory exploit a bug in dmenu when a user selects input, or mess with
the clipboard.