Contradict the last commit completely.
Well, I figured out that I could handle all of the OS checking stuff in
the module itself and realized how much effort that saves me. Checked
this on a different OS and it works. So I'm adding back in the module,
except this time it's Anthesis::OpenBSD so that it's more convenient to
call.

Only problem now is that it adds both OpenBSD::Pledge and
OpenBSD::Unveil to modules, even if only one is used.