Preserve original background setting when toggling Goyo mode
A change was made to the autocmd for 'QuitPre' in 'autocmds.lua' and the
keybinding for '<leader>g' in 'keybindings.lua'. Before Goyo mode is
activated, the current 'background' setting is stored. After Goyo is
activated, the original 'background' setting is reapplied. This prevents
Neovim's default behavior of changing the background to "dark" when
toggling Goyo mode. A relevant issue link has been added as a comment
above these changes for more context.