Change neovim plugins to be one per file
While it may seem a little ridiculous to have such small snippets in
separate files, this is for a reason. It allows for a greater level of
flexibility and consistency. I like the idea of thinking of plugins as
'modules.'

Also, by placing relevant contents into separate files, it makes it much
easier to comment or group particular plugins, and it's easier to see
what plugins I use at a glance.