chore(nvim,ksh,profile): delete website + gw stuff
This is part of an overall process to clean up these dotfiles. I won't
be developing the website the same way as before, and I don't use
OpenBSD as an OS for personal computing and development anymore. So it
doesn't make much sense for this stuff to still be here.

There's a lot of cruft that's accumulated in this repo so I expect I'll
have at least a few commits after this related to cleanup.