chore: delete things that aren't configs
My workflow is very different from how it used to be. I don't really use
any of the things I deleted anymore, and they create a lot of
complexity. By ripping these things out, it should be easier to clean up
what's left behind so that it can be more broadly useful.