fix(spectrwm): remove sysinfo bar action
This isn't a portable program, it's very tailored to OpenBSD. The
program currently doesn't check whether apm is present, so it exits
unsuccessfully. I'll either have to fix the program later or change bar
setup. Probably the second one, this is the only program written in go
within dotfiles and I'd like to make dotfiles more portable.