fix(ksh): delete user-defined completions
User-defined completions tend to be an uphill battle in OpenBSD's ksh.
Since I'm not using this shell as part of my day-to-day anymore, it may
be easier to just get rid of it. If it's annoying not to have the
completions anymore, I'll add them back in some other time. Safer to
just delete them for now though, rather than have something
unmaintained.