chore(functions): remove trailing slashes check
This check is probably bypassable and I'm not confident that it provides
any real value anyway. If anything, it probably creates more possible
situations to worry about. Also, this is a shell function, if it's used
wrong the user should be presented that error directly rather than
having the function essentially edit their input in the background.