fix(fontconfig): remove embeddedbitmap=false
This was causing waybar to act all sorts of funky when it would come
across a window title that had emojis in it. Also, I found out that if
bitmap scaling is disabled, this will make emojis gigantic in certain
situations, so I've enabled it on the system level (not in dotfiles).