fix(waybar): remove sway/window altogether
So I guess the tooltip reveals the title, so it's still getting
processed. Even if I removed the tooltip, it may not cause waybar to
ignore it, so I'm just going to remove the window module. I mainly was
using this to keep track of which window was focused and I think I can
do this a different way (window borders?).

Window borders do carry the risk of browser fingerprinting, but I'm not
as worried about that.