fix(bash): add .bash_profile file
From the man page:

When bash is invoked as an interactive login shell, or as a
non-interactive shell with the --login option, it first reads and
executes commands from the file /etc/profile, if that file exists. After
reading that file, it looks for ~/.bash_profile, ~/.bash_login, and
~/.profile, in that order, and reads and executes commands from the
first one that exists and is readable. The --noprofile option may be
used when the shell is started to inhibit this behavior.

In other words, a .bash_profile created by the system will completely
bypass sourcing the .profile provided in these dotfiles. So create one
that sources .profile to avoid this problem.