feat(emoji): update emoji file to 15.1
While here, I have to update the validation in emoji_selector as well
because otherwise the validation will omit certain lines.

Honestly, I think the validation is more of a basic assurance that bad
data won't get into the clipboard, than it is any real kind of security
measure. It's unlikely that an attacker would only have write access to
the emoji file, and much more likely that they'd have write access to
everything in the home directory. In that model, it's easy to edit the
script, or do all sorts of other things that are much nastier than just
messing with what gets put into the clipboard. So it doesn't provide any
real security.