feat(shell)!: switch over to fish
I decided that fish would be a better call over zsh, even though it's
not strictly POSIX compliant. Here are a couple of reasons:

- There's a real effort to make it friendly and usable without a ton of
  configuration. While some things in fish are possible in other POSIX
  compliant shells, they tend to require configuration, and it feels a
  bit "bolted on." I'm not against having the ability to configure
  things, but I much prefer for tools to choose good defaults in the
  first place. Neither zsh nor bash handle defaults as well as fish
  does.

- It's popular. Popularity is important, it means plenty of attention,
  bug fixes, etc.

- Fish is being rewritten in Rust, a memory safe language. It's a nice
  bonus.

This did require me to rewrite all of the aliases and functions in
fish's language, but I think that tradeoff is fine. It helped me learn
the quirks of fish a little better.

Chances are I'll still be writing scripts in POSIX shell. It's more
portable because there's an established standard for it. But for
personal + interactive use, I think it'll have to be fish. Who knows,
maybe I'll change my mind again, but I'd be surprised if it happened any
time soon. I looked at nushell and I think it's too big of a paradigm
shift to be a favorable investment of time right now, and no other
alternative shells are remotely as popular from what I can tell.