fix(profile): prepend, not append, XDG_BIN_HOME
This will be needed for bubblewrap wrapper scripts. If XDG_BIN_HOME is
appended, the wrapper script would need to be named something different,
and it's possible that I might forget to launch the wrapper instead of
the original executable, potentially creating a security problem.

Yes, it does make it so that a user writable directory's executables
come before system ones, but this is not really creating a new and
impactful security issue. An attacker that can write to ~/.local/bin can
probably also write to ~/.profile, ~/.bashrc, or whatever to execute
code on login.